pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "streamlit>=1.51.0",
]
